requirements.txt
# Production Dependencies

# Core
python-dotenv>=1.0.0

# Database & ORM
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0
pgvector>=0.2.0

# Redis & Caching
redis>=5.0.0

# LLM & AI
langchain>=0.1.0
langchain-community>=0.0.10
sentence-transformers>=2.2.0

# MCP (Model Context Protocol)
fastmcp>=0.6.0

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0

# HTTP & Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0

# Office & Data Files
openpyxl>=3.1.0
PyPDF2>=3.0.0

# Scheduling & Tasks
apscheduler>=3.10.0

# Data Science & ML
numpy>=1.24.0
scikit-learn>=1.3.0
pandas>=2.0.0

# Security & Cryptography
bcrypt>=4.1.0
PyJWT>=2.8.0
